+++ package.json
@@ -27,7 +27,8 @@
   "dependencies": {
     "@actions/core": "^1.6.0",
     "jwt-decode": "^3.1.2",
-    "node-fetch": "^2.6.7"
+    "node-fetch": "^2.6.7",
+    "graphql-request": "^4.0.0"
   },
   "devDependencies": {
     "@vercel/ncc": "^0.33.1",
